separate core
